Fix QuestionListDTO's.
